backend/requirements.txt
fastapi
uvicorn[standard]
supabase-py
python-dotenv
pydantic
rapidfuzz
openai
